Extended predictRaw method
