+++ examples/src/main/kotlin/examples/onnx/cv/efficicentnet/EfficientNet4Lite.kt
