+++ examples/src/main/kotlin/examples/onnx/cv/lenet/LenetOnMnist.kt
@@ -3,7 +3,7 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
  */
 
-package examples.onnx.mnist
+package examples.onnx.cv.lenet
 
 import org.jetbrains.kotlinx.dl.api.core.metric.Metrics
 import org.jetbrains.kotlinx.dl.api.inference.loaders.ONNXModelHub
