+++ examples/src/main/kotlin/examples/onnx/faces/Fan2D106LightAPIwithVisualization.kt
