+++ examples/src/main/kotlin/examples/onnx/objectdetection/ssdmobile/SSDMobile.kt
@@ -3,7 +3,7 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
  */
 
-package examples.onnx.objectdetection.ssd
+package examples.onnx.objectdetection.ssdmobile
 
 import examples.transferlearning.getFileFromResource
 import org.jetbrains.kotlinx.dl.api.inference.loaders.ONNXModelHub
@@ -22,7 +22,7 @@ import java.io.File
  */
 fun predictionSSD() {
     val modelHub = ONNXModelHub(cacheDirectory = File("cache/pretrainedModels"))
-    val modelType = ONNXModels.ObjectDetection.SSD
+    val modelType = ONNXModels.ObjectDetection.SSDMobileNetV1
     val model = modelHub.loadModel(modelType)
 
     model.use {
@@ -36,8 +36,8 @@ fun predictionSSD() {
                 }
                 transformImage {
                     resize {
-                        outputHeight = 1200
-                        outputWidth = 1200
+                        outputHeight = it.inputShape[1].toInt()
+                        outputWidth = it.inputShape[2].toInt()
                     }
                     convert { colorMode = ColorMode.BGR }
                 }
